refactor(mcp): split server into ServerContext + per-domain tool modules
Addresses five review findings:
- 05-mcp-monolithic-god-module: split ~1900 LOC server.py into 14
  per-domain modules under src/fabric_dw/mcp/tools/
- 05-mcp-module-singletons-race: replace _http_client / _cache_singleton
  / _resolver_singleton with a single ServerContext dataclass held in a
  module-level sentinel (_SERVER_CTX) that is set during the FastMCP
  lifespan and cleared on shutdown
- 05-mcp-no-lifecycle-shutdown: add fabric_lifespan async context manager
  that enters FabricHttpClient as an async CM, ensuring httpx connections
  are drained on SIGTERM / CTRL-C
- 05-mcp-no-di-environ-in-factories: build_context(environ) accepts a
  Mapping so tests can pass a fake env without monkey-patching os.environ
- 05-mcp-no-explicit-tool-names: every @mcp.tool() now carries an
  explicit name= keyword to prevent accidental renames

All 65 tools preserved with unchanged names/signatures. Every _guards.py
call, max_rows/row_limit cap, and host-binding guard preserved exactly.
Dedup work (SqlTarget ×20, qualified-name ×9) deferred to
refactor/mcp-tool-quality; canonical helpers stubbed in _helpers.py.

Tests: 1165 passed, ruff/ty clean.
